Refactor: moved plan module to separate file
